+++ adapters/Vungle/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Vungle Adapter for Google Mobile Ads SDK for iOS Changelog
 
-## Next Version
+## VERSION 6.3.2.2
+- Fixed a bug where the Vungle adapter would never load rewarded ads if Vungle SDK initialization failed. Now, the adapter will try to re-initialize the Vungle SDK on subsequent rewarded ad requests.
 - Now requires Google Mobile Ads SDK version 7.42.2 or higher.
 
 ## VERSION 6.3.2.1